requirements.txt
opencv-python
pytesseract
pillow
easyocr
